Add event notification for order processing hooks
For a client project we developed a plugin that sends every new order to an external api, which then further processes the order. To do that, our plugin subscribes to the Shopware_Modules_Order_SaveOrder_OrderCreated event, notified by the orders core module. For the BackendOrders plugin to be compatible with that, it should also emit that event.